requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis[asyncio]==5.0.4
pgvector==0.2.5
openai==1.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.4
